pyproject.toml
[project]
name = "lcn"
version = "0.1.0"
description = "Latent Consensus Networks for Multi-Agent Collaboration"
requires-python = ">=3.10,<3.14"
dependencies = [
    "torch>=2.0.0",
    "transformers>=5.2.0",
    "pyyaml>=6.0",
    "numpy>=1.24.0",
    "tqdm>=4.65.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.package-data]
lcn = ["configs/*.yaml"]

[tool.setuptools.packages.find]
include = ["lcn*"]
